+++ packaging/debian/validate-package-assets.sh
@@ -8,6 +8,7 @@ tmpfiles="$repo_root/packaging/linux/tmpfiles.d/dasobjectstore.conf"
 daemon_config="$repo_root/packaging/linux/etc/dasobjectstore/daemon.json"
 postinst="$repo_root/packaging/debian/postinst"
 build_deb="$repo_root/packaging/debian/build-deb.sh"
+build_rpm="$repo_root/packaging/rpm/build-rpm.sh"
 
 require_file() {
   local path="$1"
@@ -32,6 +33,7 @@ require_file "$tmpfiles"
 require_file "$daemon_config"
 require_file "$postinst"
 require_file "$build_deb"
+require_file "$build_rpm"
 
 require_text "$service" "User=dasobjectstore"
 require_text "$service" "Group=dasobjectstore"
@@ -56,8 +58,18 @@ require_text "$postinst" 'repair_managed_tree "$root"'
 require_text "$postinst" 'Managed DAS roots must be owned by $service_user:$service_group'
 
 require_text "$build_deb" "cargo build --release -p dasobjectstore-daemon"
+require_text "$build_deb" "dpkg-deb is required to build the DASObjectStore Debian package."
 require_text "$build_deb" 'target/release/dasobjectstored'
 require_text "$build_deb" 'lib/systemd/system/dasobjectstored.service'
 require_text "$build_deb" 'usr/lib/sysusers.d/dasobjectstore.conf'
 require_text "$build_deb" 'usr/lib/tmpfiles.d/dasobjectstore.conf'
 require_text "$build_deb" 'DEBIAN/postinst'
+
+require_text "$build_rpm" "rpmbuild"
+require_text "$build_rpm" "cargo build --release -p dasobjectstore-daemon"
+require_text "$build_rpm" 'target/release/dasobjectstored'
+require_text "$build_rpm" 'usr/lib/systemd/system/dasobjectstored.service'
+require_text "$build_rpm" 'usr/lib/sysusers.d/dasobjectstore.conf'
+require_text "$build_rpm" 'usr/lib/tmpfiles.d/dasobjectstore.conf'
+require_text "$build_rpm" 'systemd-sysusers /usr/lib/sysusers.d/dasobjectstore.conf'
+require_text "$build_rpm" 'systemd-tmpfiles --create /usr/lib/tmpfiles.d/dasobjectstore.conf'